First cut at library-izing patchwatcher to make it maintainable, reduce its learning curve, and make it non-Wine-specific.
This really should be done in Buildbot, but it's easier to prototype
in shell.  

Each function in the library can be called interactively from libpwdemo.sh,
which makes it a whole lot easier to learn and debug the code.

So far, it can download patches and upload results, but the
bits in the middle still need writing.

A    dashboard2.pl
A    libpatchwatcher.sh
A    libpwdemo.sh
A    get-patches2.pl